+++ example/extended_math_example.dart
@@ -22,7 +22,7 @@ void main() {
   final double det = v3.determinant();
 
   // Computes eigenvalues and eigenvectors of square matrix
-  final Map<String, MatrixBase> result = v3.eigenDecomposition();
+  final Map<num, Vector> result = v3.eigenDecomposition();
 
   final m = Matrix(<List<double>>[
     <double>[2, -1, 5],