+++ lib/src/complex_analysis/complex.dart
@@ -2,47 +2,56 @@ import 'dart:math' as m;
 
 import 'package:quiver/core.dart';
 
+import '../discrete_mathematics/general_algebraic_systems/number/double.dart';
 import '../discrete_mathematics/linear_algebra/vector/vector.dart';
+import '../mixins/copyable_mixin.dart';
 import 'exceptions/complex_exception.dart';
 
 /// Defines type for number i of complex analysis
-class Complex {
-  /// Accept [re] and [im] value in
+class Complex with CopyableMixin<Complex> {
+  /// Creates instance of complex number with [re] and [in] equals to zero
   Complex({this.re = 0, this.im = 0});
 
-  /// Creates complex number from real number
-  Complex.toComplex(this.re);
-
   /// Real part of i
-  num re = 0;
+  num re;
 
   /// Imaginary part of i
-  num im = 0;
+  num im;
 
   /// Module of this
   num get module => Vector(<num>[re, im]).length;
 
   /// Argument of this (angle of the radius Oz, where z is this)
-  num get argument => Vector(<num>[re, im]).angleBetween(Vector(<int>[1, 0]));
+  num get argument {
+    if (re >= 0) {
+      return m.atan(im / re);
+    } else {
+      return m.atan(im / re) + m.pi;
+    }
+  }
 
-  /// Add this to another [i]
-  Complex operator +(Object i) {
+  /// Add this to another [other]
+  ///
+  /// [other] can be either number or complex number.
+  Complex operator +(Object other) {
     Complex c;
-    if (i is num) {
-      c = Complex(re: re + i, im: im);
-    } else if (i is Complex) {
-      c = Complex(re: re + i.re, im: im + i.im);
+    if (other is num) {
+      c = Complex(re: re + other, im: im);
+    } else if (other is Complex) {
+      c = Complex(re: re + other.re, im: im + other.im);
     }
     return c;
   }
 
-  /// Subtract [i] from this
-  Complex operator -(Object i) {
+  /// Subtract [other] from this
+  ///
+  /// [other] can be either number or complex number.
+  Complex operator -(Object other) {
     Complex c;
-    if (i is Complex) {
-      c = this + -i;
-    } else if (i is num) {
-      c = Complex(re: re - i, im: im);
+    if (other is Complex) {
+      c = this + -other;
+    } else if (other is num) {
+      c = Complex(re: re - other, im: im);
     }
     return c;
   }
@@ -50,55 +59,100 @@ class Complex {
   /// Unary minus of this
   Complex operator -() => Complex(re: -re, im: -im);
 
-  /// Multiply this i by another [i]
-  Complex operator *(Complex i) {
-    final newRe = re * i.re + im * i.im * -1;
-    final newIm = re * i.im + im * i.re;
-    return Complex(re: newRe, im: newIm);
+  /// Multiply this i by [other]
+  ///
+  /// [other] can be either number or complex number.
+  Complex operator *(Object other) {
+    Complex c;
+    if (other is num) {
+      final newRe = re * other;
+      final newIm = im * other;
+      c = Complex(re: newRe, im: newIm);
+    } else if (other is Complex) {
+      final newRe = re * other.re - im * other.im;
+      final newIm = re * other.im + im * other.re;
+      c = Complex(re: newRe, im: newIm);
+    }
+    return c;
   }
 
-  /// Divide this i by another [i]
-  Complex operator /(Complex i) {
-    final down = m.pow(i.re, 2) + m.pow(i.im, 2);
-    final newRe = (re * i.re + im * i.im) / down;
-    final newIm = (im * i.re - re * i.im) / down;
-    return Complex(re: newRe, im: newIm);
+  /// Divide this i by [other]
+  ///
+  /// [other] can be either number or complex number.
+  Complex operator /(Object other) {
+    Complex c;
+    if (other is num) {
+      final down = m.pow(other, 2);
+      final newRe = (re * other) / down;
+      final newIm = (im * other) / down;
+      c = Complex(re: newRe, im: newIm);
+    } else if (other is Complex) {
+      c = this * other.conjugate();
+    }
+    return c;
   }
 
   /// Gets power of this by [power]
   ///
-  /// Accept only integer number in range of `-Infinite` to `Infinite` except zero.
+  /// Accept only integer number in range of `-Infinite` to `Infinite`.
   Complex pow(int power) {
     if (power == 0) {
-      return Complex.toComplex(1);
+      return Complex(re: 1);
     }
-    var tmpI = this;
+    var tmpI = copy();
 
-    // реалізувати для дробів
     for (var i = 1; i < power.abs(); i++) {
       tmpI *= this;
     }
 
     if (power < 0) {
-      tmpI = Complex.toComplex(1) / tmpI;
+      tmpI = Complex(re: 1) / tmpI;
     }
 
     return tmpI;
   }
 
-  /// Gets root of this by given [root]
+  /// Gets roots of this by given [root]
   ///
   /// [root] must not be equal to zero. Otherwise [ComplexException] will be thrown.
-  Complex rootOf(int root) {
+  List<Complex> rootsOf(int root) {
+    final result = <Complex>[];
     if (root == 0) {
       throw ComplexException('Еhe root of the zero degree does not exist!');
     }
 
-    final rootModule = m.pow(module, 1 / root);
-    final newRe = m.cos(argument / root) * rootModule;
-    final newIm = m.sin(argument / root) * rootModule;
+    // final rootModule = m.pow(module, 1 / root);
+    final rootModule = Double(module).rootOf(root).toDouble();
 
-    return Complex(re: newRe, im: newIm);
+    for (var i = 0; i < root; i++) {
+      final newRe = m.cos((argument + 2 * argument * i) / root) * rootModule;
+      final newIm = m.sin((argument + 2 * argument * i) / root) * rootModule;
+      result.add(Complex(re: newRe, im: newIm));
+    }
+
+    return result;
+  }
+
+  /// Return conjugate complex number `(1 / this)`
+  Complex conjugate() {
+    final denominator = m.pow(re, 2) + m.pow(im, 2);
+    return Complex(re: re / denominator, im: -im / denominator);
+  }
+
+  @override
+  Complex copy() => Complex(re: re, im: im);
+
+  /// Checks if this complex number have only real part
+  bool isReal() => im == 0;
+
+  /// Converts this complex number to real number if it have only real part
+  num toReal() {
+    if (isReal()) {
+      return re;
+    } else {
+      throw ComplexException(
+          'This complex number have imaginary part and cannot be converted to real number!');
+    }
   }
 
   @override