+++ lib/src/discrete_mathematics/general_algebraic_systems/number/exception/division_by_zero_exception.dart
@@ -1,0 +1,11 @@
+/// An exception that trows when performing division by zero
+class DivisionByZeroException implements Exception {
+  /// Creates instance of [DivisionByZeroException]
+  DivisionByZeroException();
+
+  /// Reason of exception
+  final String message = 'Division by zero is prihibited!';
+
+  @override
+  String toString() => '$message';
+}