+++ lib/src/discrete_mathematics/linear_algebra/matrix/diagonal_matrix.dart
@@ -33,4 +33,7 @@ class DiagonalMatrix extends SquareMatrix {
     }
     return DiagonalMatrix(newData).transpose().toDiagonalMatrix();
   }
+
+  @override
+  DiagonalMatrix copy() => DiagonalMatrix(data);
 }