+++ lib/src/discrete_mathematics/linear_algebra/vector/vector.dart
@@ -64,4 +64,7 @@ class Vector extends VectorBase {
       throw VectorException('Count of vector\'s numbers isn\'t equal!');
     }
   }
+
+  @override
+  Vector copy() => Vector(data);
 }