+++ lib/src/mixins/copyable_mixin.dart
@@ -1,0 +1,5 @@
+/// Mixins that makes objects copyable
+mixin CopyableMixin<T> {
+  /// Copies this instance
+  T copy();
+}