+++ test/extended_math_test.dart
@@ -9,15 +9,25 @@ void main() {
     <int>[0, 0, 0, 0]
   ]);
 
-  final i2 = Matrix(<List<num>>[
+  final i2 = SquareMatrix(<List<int>>[
     <int>[3, 2, 2],
     <int>[0, 3, 1],
-    <int>[0, 3, 1],
-    <int>[0, 3, 6]
+    <int>[0, 3, 1]
   ]);
 
   final v1 = Vector(<int>[6, 5, 7]);
   final v2 = Vector(<int>[6, 3, 1]);
 
-  print(i2.svd());
+  // print(i2.eigenDecomposition());
+
+  final result = CubicEquation(a: 1, b: 80, c: -20, d: 0).calculate();
+  //final result = QuadraticEquation(c: -4).calculate();
+  print(result);
+
+  // final c = Complex(re: -4);
+  // final c2 = Complex(re: -4);
+  // final r = c.rootsOf(2);
+  // print(c / c2);
+  // print(r);
+  // print(Complex(re: 2).pow(2));
 }